frontend/requirements.txt
dash
dash-bootstrap-components
python-dotenv